Prepared TicketCheckout for dynamic dropDown options
